requirements.txt
flask
flask-sqlalchemy 
werkzeug
PyPDF2
python-docx